Added style to detail info

+++ app/src/main/java/com/dog/sunshine/data/networkservice/CurrentWeatherJsonObject.kt
@@ -23,15 +23,15 @@ data class CurrentWeatherMainData(
     val sunset: Long = 0L,
     val temp: Float = 0F,
     val feels_like: Float = 0F, //This temperature parameter accounts for the human perception of weather. Unit Default: Kelvin, Metric: Celsius, Imperial: Fahrenheit
-    val pressure: Float,
+    val pressure: Int,
     val humidity: Float,
     val dew_point: Float = 0F, //Atmospheric temperature (varying according to pressure and humidity) below which water droplets begin to condense and dew can form.
-    val clouds: Float = 0F,
+    val clouds: Int = 0,
     val uvi: Float = 0F,
     val visibility: Long = 0L,
     val wind_speed: Float, //Unit Default: meter/sec, Metric: meter/sec, Imperial: miles/hour
     val wind_gust: Float = 0F, //Unit Default: meter/sec, Metric: meter/sec, Imperial: miles/hour
-    val wind_deg: Float, //Wind direction, degrees (meteorological) in format 360
+    val wind_deg: Int, //Wind direction, degrees (meteorological) in format 360
     @Embedded
     val weather: Array<Description>,
     val rain: Float = 0F,
@@ -53,15 +53,15 @@ data class Daily(
     val temp: Temp,
     @Embedded
     val feels_like: Temp,
-    val pressure: Float,
+    val pressure: Int,
     val humidity: Float,
     val dew_point: Float = 0F, //Atmospheric temperature (varying according to pressure and humidity) below which water droplets begin to condense and dew can form.
-    val clouds: Float = 0F,
+    val clouds: Int = 0,
     val uvi: Float = 0F,
     val visibility: Long = 0L,
     val wind_speed: Float, //Unit Default: meter/sec, Metric: meter/sec, Imperial: miles/hour
     val wind_gust: Float = 0F, //Unit Default: meter/sec, Metric: meter/sec, Imperial: miles/hour
-    val wind_deg: Float, //Wind direction, degrees (meteorological) in format 360
+    val wind_deg: Int, //Wind direction, degrees (meteorological) in format 360
     @Embedded
     val weather: Array<Description>,
     val rain: Float = 0F,
